==> notebooks/1_System_selection.ipynb <==
# Phase 1: System selection
## Overview
Here I will be looking through the PVDAQ solar panel metadata csv file and I will decide what solar panel systems to use.

## Initial data inspection

import pandas as pd
metadata_df = pd.read_csv('../../raw_data/systems_metadata.csv')
display(metadata_df.head())

## Filtering what systems I need
Firstly I need solar panels with a large amount of high quality data therefore the data will be filtered for 5+ years and passing quality assurance tests. Next I will look at solar panels with my desired configuration, single axis tracking and ground mounted solar panels. 

From these results, the tracking is mostly fixed and the type count is mostly unkown. This creates quite a big problem for my origional planned configuration, instead I will switching to a fixed system and use all types of solar panels. The main differences between roof and ground mounted solar panels are the tilt of the panels and scale of the farm. Using the tilt as one of my features and using specific yield should allow me to create the model without having to know the type.

value_type_counts = metadata_df['type'].value_counts()
value_tracking_counts = metadata_df['tracking'].value_counts()

print('Type count:')
for item, count in value_type_counts.items():
    print(item, count)

print('\nTracking count:')
for item, count in value_tracking_counts.items():
    print(item, count)

filtered_df = metadata_df[
    (metadata_df['years'] > 5) &
    (metadata_df['tracking'] == 'fixed') &
    (metadata_df['qa_status'] == 'pass')
]

## Cleaning the metadata
Some of the data I need for the features of the model is available in the meta data file so I will be saving it into a csv file.

I will also look through the feature data to make sure no values are missing. If they are I will remove them from the dataframe.

needed_columns = [
    'system_id', 'latitude', 'longitude', 'elevation_m', 'azimuth', 'tilt', 'dc_capacity_kW'
]
final_df = filtered_df[needed_columns].copy()

# Iterates through each row and checks for NaN values, dropping the row if there are any.
for index, row in final_df.iterrows():
    if row.isna().any():
        final_df.drop(index, inplace=True)
            
display(final_df.head())
final_df.to_csv('../../processed_data/initial_system_metadata.csv', index=False)
