==> notebooks/3_Feature_data_preprocessing_old.ipynb <==
# Phase 3: Feature data preprocessing
## Overview
In this section I will aquiring and cleaning data for the independent variables for each system. I already have: latitude, longitude, elevation_m, azimuth, tilt. I need to get data on: solar irradiance, temperature, cloud cover, wind speed, humidity, and precipitation. I'm using the NASA POWER to get this data.

## Importing needed libraries

import requests
import pandas as pd
from io import StringIO
import time

sites_df = pd.read_csv('../processed_data/useable_systems_metadata.csv')
display(sites_df.head())

## Initial data inspection
First I'm going to look at how the data I need for each system is stored in the NASA POWER.

# These the variables I want data for: solar irradiance, air temp, cloud cover, wind speed, precipitation, humidity.
params = 'ALLSKY_SFC_SW_DWN,T2M,CLOUD_AMT,WS10M,PRECTOTCORR,RH2M'

# Sets the url for the request
base_url = "https://power.larc.nasa.gov/api/temporal/daily/point?parameters=" + params + "&community=RE"
system_url = base_url + "&longitude=-93.162525&latitude=44.914573&start=20190101&end=20191231&format=CSV"
response = requests.get(system_url)
print(response.text[:1000])

## Extracting data for each system
Next I'm going to download this data for each system and store it.

# The folder each file will be stored in
base_file_path = '../raw_data/system_weather_data/'

print("Started")
for index, row in sites_df.iterrows():
    # Gets needed information
    lon = str(row['longitude'])
    lat = str(row['latitude'])
    sys_id = str(int(row['system_id']))

    # Gets the data for that systems coordinates
    system_url = base_url + "&longitude=" + lon + "&latitude=" + lat + "&start=20190101&end=20191231&format=CSV"
    response = requests.get(system_url)

    # Saves that data
    df = pd.read_csv(StringIO(response.text), skiprows=14)
    df.to_csv(base_file_path + sys_id + '.csv', index=False)

    # Prevents spamming the server
    time.sleep(1)
    # Reports progress
    if index % 100 == 0:
        print(index)
        
print("Ended")

## Final data preprocessing
Now I'm going to use all that data ive downloaded. I will also be dropping the data on capacity and generation as I won't need this for my model.

base_file_path = '../raw_data/system_weather_data/'
final_systems_metadata_df = sites_df
# The column names I'll be using
variables = ['solar irradiance', 'temperature', 'cloud cover', 'wind speed', 'humidity', 'precipitation']
# Names of the variables in the data I downloaded
variable_names = ['ALLSKY_SFC_SW_DWN', 'T2M', 'CLOUD_AMT', 'WS10M', 'PRECTOTCORR', 'RH2M']

# Creates all of the columns we need
for variable in variables:
    final_systems_metadata_df[variable] = None
      
for index, row in final_systems_metadata_df.iterrows():
    sys_id = str(int(row['system_id']))
    # Reads the data for that system
    df = pd.read_csv(base_file_path + sys_id + '.csv')

    # Finds the mean for each varibale and stores it in the df
    for i, variable in enumerate(variables):
        final_systems_metadata_df.loc[index, variable] = df[variable_names[i]].mean()

final_df = final_systems_metadata_df.drop(columns=['dc_capacity_kW', 'mean_power_generation_kW'])
display(final_df.head())

final_df = final_df.rename(columns={'azimuth': 'abs(azimuth - 180)',})

for index, row in final_df.iterrows():
    final_df.loc[index, 'abs(azimuth - 180)'] = abs(row['abs(azimuth - 180)'] - 180)

display(final_df.head())
final_df.to_csv('../processed_data/final_systems_metadata.csv', index=False)
