==> notebooks/4_Data_Partitioning.ipynb <==
# Phase 4: Data partitioning
## Overview
In this section I will be splitting the data into 3 sets: the training set, the validation set, and the testing set. I will be randomly splitting the data in the ratio: 60% training, 20% validation, and 20% test.

import pandas as pd
sites_df = pd.read_csv('../processed_data/final_systems_metadata.csv')
    
random_state = 42

# Creates the test set
df_test = sites_df.sample(frac=0.2, random_state=random_state)
df_remaining = sites_df.drop(df_test.index)

# Creates the validation set
df_val = df_remaining.sample(frac=0.25, random_state=random_state)

# Creates the training set
df_train = df_remaining.drop(df_val.index)

# Saves each set
df_test.to_csv('../processed_data/dataset_splits/test_set.csv', index=False)
df_val.to_csv('../processed_data/dataset_splits/validation_set.csv', index=False)
df_train.to_csv('../processed_data/dataset_splits/training_set.csv', index=False)
