==> notebooks/5_Model_Training.ipynb <==
# Phase 5: Training the model
## Overview
In this section I will training the regression model. I will look at the RMSE for each degree of polynomial from 1-5 then use the degree with the lowest RMSE.

## Importing needed files and libraries

import pandas as pd
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error
import numpy as np
import joblib

training_df = pd.read_csv('../processed_data/dataset_splits/training_set.csv')
validation_df = pd.read_csv('../processed_data/dataset_splits/validation_set.csv')
display(training_df.head())

## Running the model
Now we're going to run the model 5 times for each polynomial and calculate the RMSE.

# Sets the columns for the features and label for training
training_input = training_df.drop(columns=['performance_ratio'])
training_output = training_df['performance_ratio']

# Sets the columns for the features and label for validation
validation_input = validation_df.drop(columns=['performance_ratio'])
validation_output = validation_df['performance_ratio']

# The pipeline that will run the model
poly_model = Pipeline([
    ('scaler', StandardScaler()),
    ('poly', PolynomialFeatures()),
    ('linear', LinearRegression())
])


for i in range(1,6):
    # Sets the degree and fits the model with the training data
    poly_model.set_params(poly__degree=i)
    poly_model.fit(training_input, training_output)
    
    # Predicts the model output for the validation data
    predicted_validation_output = poly_model.predict(validation_input)
    
    rmse = np.sqrt(mean_squared_error(predicted_validation_output, validation_output))
    print(rmse)

## Saving the model
We can see that a degree 1 model has the lowest RMSE therefore we will be using linear regression.

# Combines the validation and training data
combined_input = pd.concat([training_input, validation_input], axis=0, ignore_index=True)
combined_output = pd.concat([training_output, validation_output], axis=0, ignore_index=True)

# Sets the final model parameters and fits it to the combined data
poly_model.set_params(poly__degree=1)
poly_model.fit(combined_input,combined_output)

# Saves the model to be loaded later
model_file = '../models/old_model.joblib'
joblib.dump(poly_model, model_file)
print("Model saved")
